refactor(my-page): 마이페이지 위젯 추출
